+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Features
 
 -   Added `BackwardIndefiniteIntegral` symbol ([#1014](https://github.com/pybamm-team/PyBaMM/pull/1014))
+-   Added `plot` and `plot2D` to enable easy plotting of `pybamm.Array` objects ([#1008](https://github.com/pybamm-team/PyBaMM/pull/1008))
 -   Added SEI film resistance as an option ([#994](https://github.com/pybamm-team/PyBaMM/pull/994))
 -   Allow symbolic inputs in initial conditions ([#968](https://github.com/pybamm-team/PyBaMM/pull/968))
 -   Added solver-specific input `extra_options` which gets passed down to each solver ([#968](https://github.com/pybamm-team/PyBaMM/pull/968))